requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cloudinary==1.32.0
dj-database-url==1.2.0
dj-email-url==1.0.6
Django==4.1.7
django-cache-url==3.4.4
django-ckeditor==6.5.1
django-js-asset==2.0.0
environs==9.5.0
gunicorn==20.1.0
marshmallow==3.19.0
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.2.0
